Create Invoice data to Excel V1
The beginning of an attempt to automate the verification of internal billing data to what was actually sent to the customer. This first iteration was utilized to verify the amounts and invoices that matched up, as well as identify the invoices that were never printed due to an internal system failure.